Add files via upload
Question 1
Write a python function that takes 4 arguments a, b, N and fn where

a and b determine an interval in 
N is a positive integer and
fn is a real valued function
The function you write should calculate the Riemann sum of the function fn over the interval [a,b] with N equal subintervals.

Solution:
Riemann Sum of a function will estimate the area under the function and the x-axis, that is, the integral of the function, between an interval. This is done by forming a number of rectangles with equal width and calculating and summing the area of these rectangles which is in fact an indicator of the area under the function. In this program, we will use the middle point of the width of each rectangle to calculate its height (f(x)), and then by multiplication of the width and height, we will obtain the area.